+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/auth/AuthAuthorityDao.java
@@ -1,6 +1,6 @@
-package guru.qa.niffler.data.dao;
+package guru.qa.niffler.data.dao.auth;
 
-import guru.qa.niffler.data.entity.AuthorityEntity;
+import guru.qa.niffler.data.entity.auth.AuthorityEntity;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
