+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/spend/CategoryDao.java
@@ -1,6 +1,6 @@
-package guru.qa.niffler.data.dao;
+package guru.qa.niffler.data.dao.spend;
 
-import guru.qa.niffler.data.entity.CategoryEntity;
+import guru.qa.niffler.data.entity.spend.CategoryEntity;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
