+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/userdata/UserdataUserDao.java
@@ -1,6 +1,6 @@
-package guru.qa.niffler.data.dao;
+package guru.qa.niffler.data.dao.userdata;
 
-import guru.qa.niffler.data.entity.UserEntity;
+import guru.qa.niffler.data.entity.userdata.UserEntity;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
