--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/Authority.java
@@ -1,5 +1,0 @@
-package guru.qa.niffler.data.entity;
-
-public enum Authority {
-  read, write
-}
