+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/auth/AuthUserEntity.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.data.entity;
+package guru.qa.niffler.data.entity.auth;
 
 import guru.qa.niffler.model.AuthUserJson;
 import java.io.Serializable;
