+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/auth/Authority.java
@@ -1,0 +1,5 @@
+package guru.qa.niffler.data.entity.auth;
+
+public enum Authority {
+  read, write
+}
