+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/spend/CategoryEntity.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.data.entity;
+package guru.qa.niffler.data.entity.spend;
 
 import guru.qa.niffler.model.CategoryJson;
 import java.io.Serializable;
