+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/spend/SpendEntity.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.data.entity;
+package guru.qa.niffler.data.entity.spend;
 
 import guru.qa.niffler.model.CurrencyValues;
 import guru.qa.niffler.model.SpendJson;
