+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/userdata/UserEntity.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.data.entity;
+package guru.qa.niffler.data.entity.userdata;
 
 import guru.qa.niffler.model.CurrencyValues;
 import guru.qa.niffler.model.UserJson;
