+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/impl/spend/CategoryDaoJdbc.java
@@ -1,9 +1,9 @@
-package guru.qa.niffler.data.impl;
+package guru.qa.niffler.data.impl.spend;
 
 import static guru.qa.niffler.data.tpl.Connections.holder;
 import guru.qa.niffler.config.Config;
-import guru.qa.niffler.data.dao.CategoryDao;
-import guru.qa.niffler.data.entity.CategoryEntity;
+import guru.qa.niffler.data.dao.spend.CategoryDao;
+import guru.qa.niffler.data.entity.spend.CategoryEntity;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
