+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/impl/spend/SpendDaoSpringJdbc.java
@@ -1,9 +1,9 @@
-package guru.qa.niffler.data.impl;
+package guru.qa.niffler.data.impl.spend;
 
 import guru.qa.niffler.config.Config;
-import guru.qa.niffler.data.dao.SpendDao;
-import guru.qa.niffler.data.entity.SpendEntity;
-import guru.qa.niffler.data.mapper.SpendEntityRowMapper;
+import guru.qa.niffler.data.dao.spend.SpendDao;
+import guru.qa.niffler.data.entity.spend.SpendEntity;
+import guru.qa.niffler.data.mapper.spend.SpendEntityRowMapper;
 import guru.qa.niffler.data.tpl.DataSources;
 import java.sql.PreparedStatement;
 import java.sql.Statement;
