+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/impl/userdata/UserdataUserDaoJdbc.java
@@ -1,9 +1,9 @@
-package guru.qa.niffler.data.impl;
+package guru.qa.niffler.data.impl.userdata;
 
 import static guru.qa.niffler.data.tpl.Connections.holder;
 import guru.qa.niffler.config.Config;
-import guru.qa.niffler.data.dao.UserdataUserDao;
-import guru.qa.niffler.data.entity.UserEntity;
+import guru.qa.niffler.data.dao.userdata.UserdataUserDao;
+import guru.qa.niffler.data.entity.userdata.UserEntity;
 import guru.qa.niffler.model.CurrencyValues;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
