+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/mapper/spend/SpendEntityRowMapper.java
@@ -1,7 +1,7 @@
-package guru.qa.niffler.data.mapper;
+package guru.qa.niffler.data.mapper.spend;
 
-import guru.qa.niffler.data.entity.CategoryEntity;
-import guru.qa.niffler.data.entity.SpendEntity;
+import guru.qa.niffler.data.entity.spend.CategoryEntity;
+import guru.qa.niffler.data.entity.spend.SpendEntity;
 import guru.qa.niffler.model.CurrencyValues;
 import java.sql.ResultSet;
 import java.sql.SQLException;
