+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/mapper/userdata/UserdataUserEntityRowMapper.java
@@ -1,6 +1,6 @@
-package guru.qa.niffler.data.mapper;
+package guru.qa.niffler.data.mapper.userdata;
 
-import guru.qa.niffler.data.entity.UserEntity;
+import guru.qa.niffler.data.entity.userdata.UserEntity;
 import guru.qa.niffler.model.CurrencyValues;
 import java.sql.ResultSet;
 import java.sql.SQLException;
