+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/AuthUserJson.java
@@ -1,6 +1,6 @@
 package guru.qa.niffler.model;
 
-import guru.qa.niffler.data.entity.AuthUserEntity;
+import guru.qa.niffler.data.entity.auth.AuthUserEntity;
 import java.util.UUID;
 import lombok.Getter;
 import lombok.Setter;
