+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/CategoryJson.java
@@ -1,7 +1,7 @@
 package guru.qa.niffler.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import guru.qa.niffler.data.entity.CategoryEntity;
+import guru.qa.niffler.data.entity.spend.CategoryEntity;
 import java.util.UUID;
 
 public record CategoryJson(
