+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/UserJson.java
@@ -1,7 +1,7 @@
 package guru.qa.niffler.model;
 
 
-import guru.qa.niffler.data.entity.UserEntity;
+import guru.qa.niffler.data.entity.userdata.UserEntity;
 import java.nio.charset.StandardCharsets;
 import java.util.UUID;
